Move to edit helper class
